==> Lessons/10_GLMs.Rmd <==
---
title: "10: Generalized Linear Models (T-Test)"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2020"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Describe the components of the generalized linear model (GLM)
2. Apply special cases of the GLM (t-test) to real datasets
3. Interpret and report the results of t-tests in publication-style formats

## Set up
```{r, message = FALSE}
getwd()
library(tidyverse)

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv")

# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## Generalized Linear Models (GLMs)

The one-sample test (model of the mean), two-sample t-test, analysis of variance (ANOVA), and linear regression are all special cases of the **generalized linear model** (GLM). The GLM also includes analyses not covered in this class, including logistic regression, multinomial regression, chi square, and log-linear models. The common characteristic of general linear models is the expression of a continuous response variable as a linear combination of the effects of categorical or continuous explanatory variables, plus an error term that expresses the random error associated with the coefficients of all explanatory variables. The explanatory variables comprise the deterministic component of the model, and the error term comprises the stochastic component of the model. Historically, artificial distinctions were made between linear models that contained categorical and continuous explanatory variables, but this distinction is no longer made. The inclusion of these models within the umbrella of the GLM allows models to fit the main effects of both categorical and continuous explanatory variables as well as their interactions. 

### Choosing a model from your data: A "cheat sheet"

**T-test:** Continuous response, one categorical explanatory variable with two categories (or comparison to a single value if a one-sample test).

**One-way ANOVA (Analysis of Variance):** Continuous response, one categorical explanatory variable with more than two categories.

**Two-way ANOVA (Analysis of Variance)** Continuous response, two categorical explanatory variables.

**Single Linear Regression** Continuous response, one continuous explanatory variable.

**Multiple Linear Regression** Continuous response, two or more continuous explanatory variables.

**ANCOVA (Analysis of Covariance)** Continuous response, categorical explanatory variable(s) and  continuous explanatory variable(s).

If multiple explanatory variables are chosen, they may be analyzed with respect to their **main effects** on the model (i.e., their separate impacts on the variance explained) or with respsect to their **interaction effects,** the effect of interacting explanatory variables on the model. 

### Assumptions of the GLM

The GLM is based on the assumption that the data residuals approximate a normal distribution (or a linearly transformed normal distribution). We will discuss the non-parametric analogues to several of these tests if the assumptions of normality are violated. For tests that analyze categorical explanatory variables, the assumption is that the variance in the response variable is equal among groups. Note: environmental data often violate the assumptions of normality and equal variance, and we will often proceed with a GLM even if these assumptions are violated. In this situation, justifying the decision to proceed with a linear model must be made. 

## T-Test
### One-sample t-test
The object of a one sample test is to test the null hypothesis that the mean of the group is equal to a specific value. For example, we might ask ourselves (from the EPA air quality processed dataset): 

Are Ozone levels below the threshold for "good" AQI index (0-50)?

```{r}

summary(EPAair$Ozone)
EPAair.subsample <- sample_n(EPAair, 5000) #take a random assortment of the 5000 samples- in this case our p value is very small so we do not have a normal distribution-- not well approximated by a normal distribution graph

# Evaluate assumption of normal distribution
shapiro.test((EPAair.subsample$Ozone)) #test can only take up 5,000 samples
ggplot(EPAair, aes(x = Ozone)) +
  geom_histogram() 
qqnorm(EPAair$Ozone); qqline(EPAair$Ozone) #Q-Q line if the data was normally distributed. There are several sampples quantiles that are higher than our one to one

O3.onesample <- t.test(EPAair$Ozone, mu = 50, alternative = "less")
O3.onesample #save this test as an object if you need to refer back to it. Just give it one column of data to use, compare to the mean (50), and an alternative of less which is the alternative hypothesis that the mean is less than 50

Ozone.plot <- ggplot(EPAair, aes(x = Ozone)) + #plot the results 
  #geom_density(stat = "count", fill = "gray") + #showing actua counts of the data set
  geom_density(fill = "gray") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) #move axis to start at 0 but must specify for both x and y
print(Ozone.plot)
```

Write a sentence or two about the results of this test. Include both the results of the test and an interpretation that puts the findings in context of the resarch question.

> It shows that the majority of our data is below 50, which is a skewed distribution of the data. The highest count appears to be 40. It shows that the ozone levels are below the threshold for "good" AQI index (0-50); however there a still counts of 100 or so of data points that lie above the threshold for ozone AQI values. We can accept the hypothesis the most of the values are below. We used a T test show statistical significance and the mean was within the confidence interval.
~~ Correct answer ~~
EOA ozone measurments for 2018-2019 were significantly lower than 50, the AQI threshold for "good" air quality (t-test: t= -57.98, p< 0.0001)

### Two-sample t-test
The two-sample *t* test is used to test the hypothesis that the mean of two samples is equivalent. Unlike the one-sample tests, a two-sample test requires a second assumption that the variance of the two groups is equivalent. Are Ozone levels different between 2018 and 2019?

```{r}
shapiro.test(EPAair$Ozone[EPAair$Year == 2018]) #run shapiro test where the year is set to 2018
shapiro.test(EPAair$Ozone[EPAair$Year == 2019]) 
var.test(EPAair$Ozone ~ EPAair$Year) #test if the variance is normal distribution, and gives us information about the alternative hypothesis- variances are significantly different from one another

ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_freqpoly() #frequency polygon to symbolize thee tests

# Format as a t-test
O3.twosample <- t.test(EPAair$Ozone ~ EPAair$Year) #compare ozone by year instead of by mu but we did not specify another null hypothesis 
O3.twosample #reject the null hypothesis because the difference is not = to 0 that the means are equal, our T value is negative and lower on our significant factor, shows the mean of each group for both years. Significant worsening in air quality. Df= degree of freedom and the higher the number the better
O3.twosample$p.value #just called up the item to find just the p value

# Format as a GLM
O3.twosample2 <- lm(EPAair$Ozone ~ EPAair$Year) #generalized lineral model (lm)
summary(O3.twosample2)
#treating year as continuous and we are expecting to se a 0.84 change per year in ozone value.
plot(O3.twosample2)
#par(mfrow= c(2,2)) to see all the tables in one image frame
```

### Non-parametric equivalent of t-test: Wilcoxon test

When we wish to avoid the assumption of normality, we can apply *distribution-free*, or non-parametric, methods in the form of the Wilcoxon rank sum (Mann-Whitney) test. The Wilcoxon test replaces the data by their rank and calculates the sum of the ranks for each group. Notice that the output of the Wilcoxon test is more limited than its parametric equivalent.

```{r}
O3.onesample.wilcox <- wilcox.test(EPAair$Ozone, mu = 50, alternative = "less")
O3.onesample.wilcox #V statisitic (not T) but still gives a P value -- less statistical robustness
O3.twosample.wilcox <- wilcox.test(EPAair$Ozone ~ EPAair$Year)
O3.twosample.wilcox #W statistics for two tests with a P-value
```

### Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year . Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#1 -- must work on completing this for next class minus adding letters representing with key
plot.1<-
  ggplot(EPAair, aes(x = Ozone, color = as.factor(Year), alpha =0.2)) +
  geom_density(fill="gray")+
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) 
plot.1

#2
plot.2 <-
  ggplot(EPAair, aes(x = Ozone, color = as.factor(Year), alpha =0.2)) +
  geom_boxplot()+
  #stat_summary()
plot.2

#3
plot.3<-
  ggplot(EPAair, aes(x = Ozone, color = as.factor(Year), alpha =0.2)) +
  geom_violin()%>%
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9)%>%
plot.3
```

Now, write a summary of your findings, incorporating statistical output, reference to the figure(s), and a contextual interpretation.

>
